Separate report managers into partial classes
