+++ src/Magenta.Workflow/Managers/Reports/FlowIdentityReportManager.cs
@@ -1,0 +1,7 @@
+﻿namespace Magenta.Workflow.Managers.Reports
+{
+    public partial class FlowReportManager
+    {
+
+    }
+}
