+++ src/Magenta.Workflow/Managers/Reports/FlowInboxReportManager.cs
@@ -1,0 +1,7 @@
+﻿namespace Magenta.Workflow.Managers.Reports
+{
+    public partial class FlowReportManager
+    {
+        //Inbox reports
+    }
+}
