+++ src/Magenta.Workflow/Managers/Reports/FlowOverviewReportManager.cs
@@ -1,0 +1,10 @@
+﻿namespace Magenta.Workflow.Managers.Reports
+{
+    public partial class FlowReportManager
+    {
+
+        // Number of instances, types, users, roles(groups)
+        //Current status of flow
+
+    }
+}
